requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
transformers>=4.30.0
torch>=2.2.0
ollama>=0.1.7
gitpython>=3.1.40
bandit>=1.7.5
safety>=2.3.5
pylint>=3.0.3
black>=23.11.0
isort>=5.12.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
jinja2>=3.1.2
pyyaml>=6.0.1
